+++ proto/common/common.proto
@@ -63,9 +63,23 @@ message Chain {
   int64 chain_id = 2;
 }
 
+message BlockHeader {
+  int64 height = 1;
+  bytes hash = 2;
+  bytes parent_hash = 3;
+  int64 chain_id = 4;
+  // chain specific header
+  HeaderData header = 5 [(gogoproto.nullable) = false];
+}
+
+message HeaderData {
+  oneof data {
+    bytes ethereum_header = 1; // binary encoded headers; RLP for ethereum
+  }
+}
+
 message Proof {
-  int64 chain_id = 1;
   oneof proof {
-    ethereum.Proof eth_proof = 10;
+    ethereum.Proof ethereum_proof = 1;
   }
 }
