--- proto/observer/headers.proto
@@ -1,22 +1,0 @@
-syntax = "proto3";
-package zetachain.zetacore.observer;
-
-import "gogoproto/gogo.proto";
-import "observer/observer.proto";
-
-option go_package = "github.com/zeta-chain/zetacore/x/observer/types";
-
-message BlockHeader {
-  int64 height = 1;
-  bytes hash = 2;
-  bytes parent_hash = 3;
-  int64 chain_id = 4;
-  // chain specific header
-  HeaderData header = 5 [(gogoproto.nullable) = false];
-}
-
-message HeaderData {
-  oneof data {
-    bytes ethereum_header = 1; // binary encoded headers; RLP for ethereum
-  }
-}
