+++ x/crosschain/types/expected_keepers.go
@@ -66,7 +66,7 @@ type ZetaObserverKeeper interface {
 	IsAuthorized(ctx sdk.Context, address string, chain *common.Chain) (bool, error)
 	FindBallot(ctx sdk.Context, index string, chain *common.Chain, observationType zetaObserverTypes.ObservationType) (ballot zetaObserverTypes.Ballot, isNew bool, err error)
 	AddBallotToList(ctx sdk.Context, ballot zetaObserverTypes.Ballot)
-	GetBlockHeader(ctx sdk.Context, hash []byte) (val zetaObserverTypes.BlockHeader, found bool)
+	GetBlockHeader(ctx sdk.Context, hash []byte) (val common.BlockHeader, found bool)
 }
 
 type FungibleKeeper interface {
