+++ x/observer/keeper/block_header.go
@@ -6,20 +6,21 @@ import (
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
+	"github.com/zeta-chain/zetacore/common"
 	"github.com/zeta-chain/zetacore/x/observer/types"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
 // SetBlockHeader set a specific block header in the store from its index
-func (k Keeper) SetBlockHeader(ctx sdk.Context, header types.BlockHeader) {
+func (k Keeper) SetBlockHeader(ctx sdk.Context, header common.BlockHeader) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.BlockHeaderKey))
 	b := k.cdc.MustMarshal(&header)
 	store.Set(header.Hash, b)
 }
 
 // GetBlockHeader returns a block header from its hash
-func (k Keeper) GetBlockHeader(ctx sdk.Context, hash []byte) (val types.BlockHeader, found bool) {
+func (k Keeper) GetBlockHeader(ctx sdk.Context, hash []byte) (val common.BlockHeader, found bool) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.BlockHeaderKey))
 
 	b := store.Get(hash)
@@ -46,9 +47,9 @@ func (k Keeper) GetAllBlockHeaders(c context.Context, req *types.QueryAllBlockHe
 	store := ctx.KVStore(k.storeKey)
 	blockHeaderStore := prefix.NewStore(store, types.KeyPrefix(types.BlockHeaderKey))
 
-	var blockHeaders []*types.BlockHeader
+	var blockHeaders []*common.BlockHeader
 	pageRes, err := query.Paginate(blockHeaderStore, req.Pagination, func(key []byte, value []byte) error {
-		var blockHeader types.BlockHeader
+		var blockHeader common.BlockHeader
 		if err := k.cdc.Unmarshal(value, &blockHeader); err != nil {
 			return err
 		}
